requirements.txt
rapidocr-onnxruntime>=1.3.0
python-doctr[torch]>=0.7.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
SQLAlchemy>=2.0.0
watchdog>=3.0.0
numpy>=1.24.0
shapely>=2.0.0